requirements.txt
streamlit
load_dotenv
openai